Fix all 13 code-review defects with per-finding RED regression tests;
134/134 green in both the default and LibASan (ZODIAC_INSTRUMENT_LIB=ON)
builds.

+++ src/z_cfile.h
@@ -47,8 +47,8 @@ protected:
 
 	FILE * file{};
 	bool   ownsFile{};
-	uint8_t stackSize{};
-	uint8_t stackPos{};
+	uint32_t stackSize{};
+	uint32_t stackPos{};
 	StackFrame * stack{};
 };
